test: move proxy cache-eligibility + survival tests to test_proxy.py
Per Goykhman's r3277892192: these tests are logically test_proxy
material since the consolidation from @cres_cacheable → @proxy.
Moved from test/utils/test_highlevel.py to test/core/test_proxy.py:

- _sole_compile_result() helper
- test_proxy_zero_arg_caller_is_cacheable (errno_get caller)
- test_proxy_single_arg_caller_is_cacheable (getenv caller)
- test_proxy_multi_arg_caller_is_cacheable (memcpy caller)
- test_proxy_caller_survives_subprocess_round_trip

test_orphan_anchor_sweep_spares_fresh_tmp_files stays in
test_highlevel.py — _orphan_anchor_sweep is in highlevel.py for
make_structref.

Import cleanup:
- test_highlevel.py drops errno_get, getenv, memcpy, array_data_p,
  get_unicode_data_p, assert_njit_cache_survives_subprocess_roundtrip,
  numpy as np (only `from numpy import isclose` still needed).
- test_proxy.py gains the same set plus njit.

402 pytest pass on Py3.12 — same count as before (test moved, not
removed); strict flake8 clean.
